pyproject.toml
[tool.poetry]
name = "db-2026"
version = "0.1.0"
description = ""
authors = ["Piotr Marecki <pmarecki@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
pydantic = "^2.12.5"
asyncpg = "^0.31.0"
python-dotenv = "^1.2.2"
faker = "^40.13.0"
loguru = "^0.7.3"


[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"
pytest-asyncio = "^1.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
